refactor(be): centralize artiment mention config
